renamed package to serverless-dev-runtime
