+++ packages/serverless-dev-runtime/lib/templates/default-function-package.json
